Document modifier and types
